Fixed some bitshift width warnings.

+++ libraries/lps/include/mcrl2/lps/binary.h
@@ -94,7 +94,7 @@ class binary_algorithm: public detail::lps_algorithm<Specification>
       else
       {
         std::size_t n = enumerated_elements.size();
-        std::size_t m = 1 << (new_parameters.size() - 1);  //m == 2^(new_parameters.size() - 1)
+        std::size_t m = static_cast<ssize_t>(1) << (new_parameters.size() - 1);  //m == 2^(new_parameters.size() - 1)
 
         if (m > n)
         {
@@ -226,7 +226,7 @@ class binary_algorithm: public detail::lps_algorithm<Specification>
             while (k != elements.end())
             {
               // Elements that get boolean value false
-              ssize_t count(1 << j);
+              ssize_t count(static_cast<ssize_t>(1) << j);
               if (std::distance(k, elements.end()) < count)
               {
                 k = elements.end();
